Account for entirely deleted or added nodes
Previously if whole blocks were deleted or added we handled it, but
not if nodes within a block were added or deleted (e.g. adding or deleting
a link or paragraph). This provides functionality for those scenarios, and
introduces testing the expected behaviour so we can ensure we cover all the
edge cases